+++ CHANGELOG.md
@@ -14,7 +14,7 @@
 
 ## Compatibility Notes
 
--
+- [core] Tightened the return type of `getAuthHeaders()` and `buildAuthorizationHeaders()` methods.
 
 ## New Functionality
 
@@ -46,6 +46,8 @@ Blog: TBD<br>
 ## Fixed Issues
 
 - [core] Fix a type error, when `moment()` is used in filtering an OData field of type `Edm.DateTimeOffset`
+- [core] Fix missing `Proxy-Authorization` header if custom authorization headers are set.
+- [core] Remove wrongly given warning that custom headers are given and will overwrite destination headers.
 
 # 1.49.0